[environment] Do not use pointers from out-of-scope objects. MER#1749
Doing pointer = bytearray + cstring operation constructs temporary object
that then goes out of scope after assignment, which most likely leaves the
pointer pointing to already freed and likely to be reused for something
else area in the heap.

Append the defined in the same scope QByteArray object and use the data
held within it.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>